Refactoring to produce a pure lib for cookie verification
The existing libraries are coupled to a overhead like reloading the
secrets from an S3 bucket. This refactor pulls some common
functionality out into a new pure library for cookie
verification. This is possible now that the cookie is signed
asymmetrically and can be verified just using the public key.

The next step will be to remove some of the logic duplication.