Introduce legacy cookie support for phased release
This means systems can be updated one by one, rather than us having to
coordinate the rollout.

In this commit, both cookies are dropped (which allows services to
begin doing verification on the new cookie) but the check is still
performed against the existing cookie.

Subsequent versions will:

1. perform the check against the new cookie
2. deprecate the legacy cookie and use the new cookie value on both cookie names
3. remove the temporary new cookie

Each step will need a full deploy of every product, but taking this
approach means there will be no interruption of service.