Fix pan domain settings refresher
The actor system was instantiated for each controller. This changed when
the actor system was passed down as a parameter, which in turn created
more problems as the domain settings refresher was instantiated for
each controller leading to a name collision.

This attempts to fix it by extracting the refresher logic out of the
trait, and requiring it to be injected in each controller