This fixes a couple of problems with the validation logic
We move auth validation logic into one place and add tests to address
the following issues:

1. the validation check was being called even if cacheValidation was set
2. users in the grace period skipped the validation checks

We now consistently apply the project-provided validateUser
function. This may cause problems for existing applications but only
in surprising cases and if the projects were already doing crazy
things.

The second problem was a potential security issue, although exploiting
it would depend on finding a project that behaves in a way we'd
probably class as "not recommended".

If the project-supplied validateUser function is impure then it's
possible that it would being to fail at some point after the original
session has been issued. For example, there might be a
session-revocation feature that returns false if the user has been
blacklisted in some way. This check would have been bypassed during
the grace period providing an attacker with a window of opportunity to
exploit a compromised session.

It's worth noting that while it's unlikely this affects any of our
projects, the pan-domain-auth library was designed to support these
use cases (this is why the `cache validation` feature
exists). Ignoring questions about whether these features make any
sense the way they are implemented in panda, this Pull Request fixes
this edge case.

Co-authored-by: Adam Fisher <adam.fisher@guardian.co.uk>