Remove cross builds.
Refactors in #48 introduced code that no longer compiled on the Scala 2.11 build due to the lack of right-based Either and converting a Scala lambda with no arguments to a Java Runnable.

All of our apps using the Play 2.6 integration are already on Scala 2.12 and the only app using the verification library that is still on 2.11 is S3 Uploader which is a straight-forward upgrade.

So we decided to remove the cross build.