Use custom cookies to perform anti-forgery
Previously, pan-domain auth used the Play Session to store an
anti-forgery token and the url to return to after authenticating. This
session is a cookie that will default to SameSite=Lax, and is designed
to be provided by the browser in the request to the oauth callback.
Because the request was initiated by the Google Oauth endpoint, which is
a 3rd party, the session cookie should not be returned to the app. (It
seems however that Chrome relaxes the restriction to allow this auth
flow to continue to work until the app is updated to set SameSite=None;
see <https://groups.google.com/a/chromium.org/g/blink-dev/c/AknSSyQTGYs/m/wRWG65IWBQAJ>. Firefox has no such exemption.)

So in order for the oauth callback to receive the anti-forgery token and
destination url, they must be placed on cookie(s) with SameSite=None.
The SameSite value for the play session must be configured on a per-app
basis; this PR adds the values onto custom cookies to avoid requiring
consumers to reconfigure the play session.