Fix: Lazily evaluate acceptedPublicKeys
If the `acceptedPublicKeys` val is evaluated eagerly within the
`Verification` trait then the `activePublicKey` and `alsoAccepted` vals
won't have been initialised yet from the class which extends the trait.

This was leading to `acceptedPublicKeys` evaluating to 'null' at
runtime.

The added test was failing with the pre-existing code with a very
similar message to the one observed in CODE, so we have some
confidence that it's testing for this case effectively.

See: https://docs.scala-lang.org/tutorials/FAQ/initialization-order.html

Co-authored-by: Roberto Tyley <52038+rtyley@users.noreply.github.com>